requirements.txt
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
anything2image==1.1.4
appdirs==1.4.4
apptools==5.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
audioread==3.0.1
av==10.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
boto3==1.28.50
botocore==1.31.50
bravado==11.0.3
bravado-core==6.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
configobj==5.0.8
contourpy==1.1.0
croniter==1.3.15
cycler==0.11.0
dask==2023.6.1
datasets==2.14.6
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
deepdiff==6.3.0
defusedxml==0.7.1
diffusers==0.17.1
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.10
einops==0.6.1
envisage==7.0.3
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.17.1
ffmpy==0.3.0
filelock==3.12.2
fire==0.5.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.31
gradio==3.35.2
gradio-client==0.2.7
h11==0.14.0
h5py==3.9.0
hickle==5.0.2
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
info-nce-pytorch==0.1.4
inquirer==3.1.3
iopath==0.1.10
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
Js2Py==0.74
json5==0.9.14
jsonlines==4.0.0
jsonpickle==3.0.2
jsonref==1.1.0
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter-server==2.6.0
jupyter-server-terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.23.0
kiwisolver==1.4.4
lazy-loader==0.3
librosa==0.10.1
lightgbm==4.1.0
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
linkify-it-py==2.0.2
llvmlite==0.41.1
lmdb==1.4.1
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mayavi==4.8.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
missingno==0.5.2
mistune==2.0.5
monotonic==1.6
msgpack==1.0.5
msgpack-numpy==0.4.8
multidict==6.0.4
multiprocess==0.70.15
munch==2.5.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
neptune-client==1.8.2
neptune-contrib==0.28.1
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook-shim==0.2.3
npy-append-array==0.9.16
numba==0.58.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.8.0.74
ordered-set==4.1.0
orjson==3.9.1
overrides==7.3.1
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parameterized==0.9.0
parso==0.8.3
partd==1.4.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipwin==0.5.2
platformdirs==3.6.0
plotly==5.15.0
pooch==1.7.0
portalocker==2.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.0
pybind11==2.11.1
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
pyface==8.0.0
Pygments==2.15.1
pyjsparser==2.7.1
PyJWT==2.7.0
pynndescent==0.5.10
pyparsing==3.1.0
PyPrind==2.11.3
pyrsistent==0.19.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.3
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
readchar==4.0.5
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
s3transfer==0.6.2
sacred==0.8.4
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentry-sdk==1.26.0
setproctitle==1.3.2
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
swagger-spec-validator==3.0.3
tabulate==0.9.0
tbb==2021.10.0
tenacity==8.2.2
tensorboardX==2.6.2.2
tensorpack==0.11
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.7
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.11.4
torchvision==0.14.0
torchxrayvision==1.1.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.4.1
traitsui==8.0.0
transformers==4.30.2
types-regex==2023.6.3.0
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
umap-learn==0.5.4
urllib3==1.26.6
uvicorn==0.22.0
vtk==9.2.6
wandb==0.15.4
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.0
websockets==11.0.3
wget==3.2
wrapt==1.15.0
xgboost==1.7.6
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
